fix: pop stash from inputfield-useref branch which refactored card and initial searchbar
