+++ src/components/card/types/CardType.ts
@@ -1,9 +1,17 @@
-import React from "react";
 
+
+/*
 type CardType = {
     media?: React.ReactNode, // image or video
     title: string,
     description: string
+}
+*/
+
+// type for Card, item returned by map function is the same type as the api data.
+
+export type CardType = {
+  item: ApiDataType
 }
 
 export interface ApiDataType  {
@@ -12,4 +20,3 @@ export interface ApiDataType  {
     image_url: string;
   };
 
-export default CardType;
